social_media_api/requirements.txt
django
djangorestframework
django-filter
Pillow
psycopg2
